Relocate BigTable uploader to ledger/ crate
