+++ ledger/Cargo.toml
@@ -17,6 +17,8 @@ dlopen_derive = "0.1.4"
 dlopen = "0.1.8"
 ed25519-dalek = "1.0.0-pre.4"
 fs_extra = "1.1.0"
+futures = "0.3.5"
+futures-util = "0.3.5"
 itertools = "0.9.0"
 lazy_static = "1.4.0"
 libc = "0.2.72"
@@ -40,9 +42,11 @@ solana-rayon-threadlimit = { path = "../rayon-threadlimit", version = "1.4.0" }
 solana-runtime = { path = "../runtime", version = "1.4.0" }
 solana-sdk = { path = "../sdk", version = "1.4.0" }
 solana-stake-program = { path = "../programs/stake", version = "1.4.0" }
+solana-storage-bigtable = { path = "../storage-bigtable", version = "1.4.0" }
 solana-vote-program = { path = "../programs/vote", version = "1.4.0" }
 tempfile = "3.1.0"
 thiserror = "1.0"
+tokio = { version = "0.2.22", features = ["full"] }
 trees = "0.2.1"
 
 [dependencies.rocksdb]
